+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Coverage Upload Action Changelog
 
+## 3.0.0 (June 29, 2022)
+
+- Parse all paths and display names from a config file
+- Modify post request to send a path and displayName to a remote endpoint
+- Modify post request to send data to a remote endpoint as a list of pull request objects
+
 ## 2.0.0 (May 18, 2022)
 
 - Modified action to receive body from the payload